requirements.txt
gradio
paddleocr
click8